+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Zenti",
+  "name": "zenti",
   "version": "0.0.1",
   "private": true,
   "scripts": {
